% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{accuracy_rate}
\alias{accuracy_rate}
\title{Accuracy rate.}
\usage{
accuracy_rate(solution, submission)
}
\arguments{
\item{solution}{.}

\item{submission}{.}
}
\value{
accuracy rate.
}
\description{
Accuracy rate.
}
\examples{

}
