% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{bearing_change_fix_cpp}
\alias{bearing_change_fix_cpp}
\title{Fixing the bearing changes close to 360 or -360.}
\usage{
bearing_change_fix_cpp(bearing_change)
}
\arguments{
\item{bearing_change}{bearing change vector.}
}
\value{
fixed bearing changes.
}
\description{
Fixing the bearing changes close to 360 or -360.
}
\examples{

bearing_change_fix_cpp(c(1,30,90,190,259,350))
}
