% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{brake_calc}
\alias{brake_calc}
\title{brake stat calculation R version.}
\usage{
brake_calc(speed, lag = 1)
}
\arguments{
\item{speed}{.}

\item{lag}{.}
}
\value{
speed before brake.
}
\description{
brake stat calculation R version.
}
\examples{

}
