% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{colSds}
\alias{colSds}
\title{Columnwise standard deviations.}
\usage{
colSds(X, norm_type = 0L)
}
\arguments{
\item{X}{a matrix.}

\item{norm_type}{normalization type, integer input, default 0.}
}
\value{
columnwise standard deviations.
}
\description{
Columnwise standard deviations.
}
\examples{

colSds(matrix(1:4,nrow=2))
}
