% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{dist2d}
\alias{dist2d}
\title{Distance from a point to a line linked by two other points.}
\usage{
dist2d(a, b, c)
}
\arguments{
\item{a}{.}

\item{b}{.}

\item{c}{.}
}
\value{
read-valued distance.
}
\description{
Distance from a point to a line linked by two other points.
}
\examples{

}
