% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dist2dCPP}
\alias{dist2dCPP}
\title{Distance from a point to a line linked by two other points C++ version.}
\usage{
dist2dCPP(a, b, c)
}
\arguments{
\item{a}{a numeric vetor with length 2.}

\item{b}{a numeric vetor with length 2.}

\item{c}{a numeric vetor with length 2.}
}
\value{
real-valued Eucleadian distance from point a to the line linking b and c.
}
\description{
Distance from a point to a line linked by two other points C++ version.
}
\examples{

dist2dCPP(c(1,1),c(0,0),c(0,1))
}
