% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{GeneratePolygon}
\alias{GeneratePolygon}
\title{Generate polypon points for Ramer–Douglas–Peucker algorithm visualization.}
\usage{
GeneratePolygon(pt1, pt2, dist = 1)
}
\arguments{
\item{pt1}{.}

\item{pt2}{.}

\item{dist}{.}
}
\value{
a matrix with dimension 4*2 denoting the 4 vertices of the polygon.
}
\description{
Generate polypon points for Ramer–Douglas–Peucker algorithm visualization.
}
\examples{
library(ggplot2)
point_set <- rbind(c(0,0),c(2,2))
GeneratePolygon(point_set[1,],point_set[2,])
df <- data.frame(t(GeneratePolygon(point_set[1,],point_set[2,],dist=0.2)))
ggplot(df,aes(X1,X2)) + geom_polygon(fill='lightblue',alpha=0.5)
}
