% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{LogLoss}
\alias{LogLoss}
\title{Log loss score for competition.}
\usage{
LogLoss(solution, submission, adj = 1e-15)
}
\arguments{
\item{solution}{.}

\item{submission}{.}

\item{adj}{.}
}
\value{
multinomial log loss score.
}
\description{
Log loss score for competition.
}
\examples{

}
