% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{LogLoss_Analysis}
\alias{LogLoss_Analysis}
\title{Individual scores at each record.}
\usage{
LogLoss_Analysis(solution, submission, adj = 1e-15)
}
\arguments{
\item{solution}{.}

\item{submission}{.}

\item{adj}{.}
}
\value{
multinomial log loss score.
}
\description{
Individual scores at each record.
}
\examples{

}
