% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{number_change}
\alias{number_change}
\title{Total number of number changes in a numeric vector.}
\usage{
number_change(numbers)
}
\arguments{
\item{numbers}{numeric vector.}
}
\value{
number of changes of numbers.
}
\description{
Total number of number changes in a numeric vector.
}
\examples{

number_change(c(1,1,1,2,2,3,3,5))
}
