% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{prob_to_rank}
\alias{prob_to_rank}
\title{Convert probablities to rank.}
\usage{
prob_to_rank(submission)
}
\arguments{
\item{submission}{.}
}
\value{
rank of submission.
}
\description{
Convert probablities to rank.
}
\examples{

}
