% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{prob_to_vote}
\alias{prob_to_vote}
\title{Convert probablities to votes.}
\usage{
prob_to_vote(submission)
}
\arguments{
\item{submission}{.}
}
\value{
votes of submission.
}
\description{
Convert probablities to votes.
}
\examples{

}
