% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{RDP}
\alias{RDP}
\title{Ramer–Douglas–Peucker algorithm.}
\usage{
RDP(points, epsilon)
}
\arguments{
\item{points}{.}

\item{epsilon}{.}
}
\value{
RDP points.
}
\description{
Ramer–Douglas–Peucker algorithm.
}
\examples{

}
