% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{RDPCPP}
\alias{RDPCPP}
\title{Ramer–Douglas–Peucker algorithm C++ version.}
\usage{
RDPCPP(points, epsilon)
}
\arguments{
\item{points}{point matrix with 2 columns (x and y coordinate like).}

\item{epsilon}{cut-off distance for the algorithm.}
}
\value{
RDP points.
}
\description{
Ramer–Douglas–Peucker algorithm C++ version.
}
\examples{

RDPCPP(cbind(c(1:4),c(1:4)),0.5)

# pick a trip from driver Alexander
sample_trip <- as.matrix(subset(tele_data,{trip_id==tele_data$trip_id[1]})[,c("lat","long")])
RDPCPP(sample_trip,0.001)


}
