% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rowSds}
\alias{rowSds}
\title{Rowwise standard deviations.}
\usage{
rowSds(X, norm_type = 0L)
}
\arguments{
\item{X}{a matrix.}

\item{norm_type}{normalization type, integer input, default 0.}
}
\value{
rowwise standard deviations.
}
\description{
Rowwise standard deviations.
}
\examples{

rowSds(matrix(1:4,nrow=2))
}
