% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sign_change}
\alias{sign_change}
\title{Total number of sign changes in a logical vector.}
\usage{
sign_change(signs)
}
\arguments{
\item{signs}{logical vector indicating the signs.}
}
\value{
number of changes of signs.
}
\description{
Total number of sign changes in a logical vector.
}
\examples{

sign_change(c(TRUE,FALSE,TRUE,TRUE))
}
