% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{start_calc}
\alias{start_calc}
\title{start stat calculation R version.}
\usage{
start_calc(speed, lag = 1)
}
\arguments{
\item{speed}{.}

\item{lag}{.}
}
\value{
speed after start.
}
\description{
start stat calculation R version.
}
\examples{

}
