% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tele_data}
\alias{tele_data}
\title{subset of 100 trips of data from driver Alexander.}
\format{A data frame with 85342 rows and 12 variables:
\describe{
  \item{time}{time in each trip, in second}
  \item{delta_time}{time increment in each trip, in second}
  ...
}}
\source{
\url{https://myconnections.lmig.com/docs/DOC-64395}
}
\usage{
tele_data
}
\description{
A dataset containing the prices and other attributes of almost 54,000
diamonds.
}
\keyword{datasets}
