% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{trip_match_cpp}
\alias{trip_match_cpp}
\title{trip match algorithm that calculate matching distances.}
\usage{
trip_match_cpp(M1, M2, dist_cut = 1e-04, heading_cut = 0.005)
}
\arguments{
\item{M1}{distance and heading matrix of simplified trip 1.}

\item{M2}{distance and heading matrix of simplified trip 2.}

\item{dist_cut}{distance cut-off.}

\item{heading_cut}{heading cut-off.}
}
\value{
distances of matching trip segments.
}
\description{
trip match algorithm that calculate matching distances.
}
\examples{

}
