% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{trip_match_v1}
\alias{trip_match_v1}
\title{trip matching algorithm R version 1.}
\usage{
trip_match_v1(M1, M2, dist_cut = 1e-05, heading_cut = 1e-05)
}
\arguments{
\item{M1}{.}

\item{M2}{.}

\item{dist_cut}{.}

\item{heading_cut}{.}
}
\value{
logical vector indicating matched index.
}
\description{
trip matching algorithm R version 1.
}
\examples{

}
