% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{trip_match_v3}
\alias{trip_match_v3}
\title{trip matching algorithm R version 3.}
\usage{
trip_match_v3(M1, M2, dist_cut = 0.001, heading_cut = 0.01,
  match_n_cut = 5)
}
\arguments{
\item{M1}{.}

\item{M2}{.}

\item{dist_cut}{.}

\item{heading_cut}{.}

\item{match_n_cut}{.}
}
\value{
logical vector indicating matched index.
}
\description{
trip matching algorithm R version 3.
}
\examples{

}
